requirements.txt
torch
torchvision
matplotlib
cython
opencv-python
pillow
pycocotools
matplotlib 
gdown